- Moved minor functions from main.cpp to addons.cpp, - Moved macros from main.cpp to globals, - Merged closeSDL() into handleExitState(), - Commented out unused code, - Update comments.
