+++ main.h
@@ -1,12 +1,13 @@
 #ifndef MAIN_H
-#define MAIN_H // GLOBALS.H
+#define MAIN_H // MAIN.H "GLOBALS.H"
 
 #include "SDL.h"
 #include "SDL_image.h"
 #include "SDL_mixer.h"
 
-extern SDL_Window *gWindow;
-extern SDL_Surface *gScreen;
+#define WINDOW_WIDTH 480
+#define WINDOW_HEIGHT 320
+#define FPS 15
 
 void initVideo(int window_width, int window_height);
 void initMixer();
@@ -17,13 +18,16 @@ struct WindowDimensions{
     int xPosi;
     int yPosi;
 };
-/*
-// Extern Uint32 variables
+
+// Declaring certain variables as extern.
+extern SDL_Window *gWindow;
+extern SDL_Surface *gScreen;
+
 extern Uint32 gPink;
 extern Uint32 gRed;
 extern Uint32 gBeige;
 extern Uint32 gBlue;
 extern Uint32 gDarkblue;
 extern Uint32 gDarkgreen;
-*/
+
 #endif // MAIN_H
